requirements.txt
disnake
pymongo
pillow
httpx